frontend/myWedoo-app/src/app/components/dialog/dialog.component.ts
import { Component, Inject } from '@angular/core';
import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
import { Amount } from 'src/app/models/amount';

@Component({
  selector: 'app-dialog',
  templateUrl: './dialog.component.html',
  styleUrls: ['./dialog.component.scss']
})
export class DialogComponent {

  possibleAmount: Amount;

  constructor(
    public dialogRef: MatDialogRef<DialogComponent>,
    @Inject(MAT_DIALOG_DATA) public data: any,
  ) {
    this.possibleAmount = data.amount;
  }

  onNoClick(): void {
    this.dialogRef.close();
  }

}
